refactor(onboarding): 온보딩 도메인(entity, repository, enum) common-core로 파일 위치 변경
