+++ Order-Core/src/main/java/com/goormgb/be/ordercore/onboarding/dto/response/OnboardingPreferenceGetResponse.java
@@ -2,16 +2,16 @@ package com.goormgb.be.ordercore.onboarding.dto.response;
 
 import java.util.List;
 
+import com.goormgb.be.onboarding.entity.OnboardingPreference;
 import com.goormgb.be.ordercore.onboarding.dto.OnboardingPreferenceDto;
-import com.goormgb.be.ordercore.onboarding.entity.OnboardingPreference;
 
 public record OnboardingPreferenceGetResponse(
-		List<OnboardingPreferenceDto> preferences
+	List<OnboardingPreferenceDto> preferences
 ) {
 	public static OnboardingPreferenceGetResponse from(List<OnboardingPreference> entities) {
 		List<OnboardingPreferenceDto> preferences = entities.stream()
-				.map(OnboardingPreferenceDto::from)
-				.toList();
+			.map(OnboardingPreferenceDto::from)
+			.toList();
 
 		return new OnboardingPreferenceGetResponse(preferences);
 	}
