+++ Order-Core/src/main/java/com/goormgb/be/ordercore/onboarding/service/OnboardingPreferenceService.java
@@ -11,17 +11,17 @@ import org.springframework.transaction.annotation.Transactional;
 import com.goormgb.be.global.exception.CustomException;
 import com.goormgb.be.global.exception.ErrorCode;
 import com.goormgb.be.global.support.Preconditions;
+import com.goormgb.be.onboarding.entity.OnboardingPreference;
+import com.goormgb.be.onboarding.enums.SeatHeight;
+import com.goormgb.be.onboarding.enums.Section;
+import com.goormgb.be.onboarding.enums.Viewpoint;
+import com.goormgb.be.onboarding.repository.OnboardingPreferenceRepository;
 import com.goormgb.be.ordercore.onboarding.dto.OnboardingPreferenceDto;
 import com.goormgb.be.ordercore.onboarding.dto.request.OnboardingPreferenceCreateRequest;
 import com.goormgb.be.ordercore.onboarding.dto.request.OnboardingPreferenceUpdateRequest;
 import com.goormgb.be.ordercore.onboarding.dto.response.OnboardingPreferenceCreateResponse;
 import com.goormgb.be.ordercore.onboarding.dto.response.OnboardingPreferenceGetResponse;
 import com.goormgb.be.ordercore.onboarding.dto.response.OnboardingStatusGetResponse;
-import com.goormgb.be.ordercore.onboarding.entity.OnboardingPreference;
-import com.goormgb.be.ordercore.onboarding.enums.SeatHeight;
-import com.goormgb.be.ordercore.onboarding.enums.Section;
-import com.goormgb.be.ordercore.onboarding.enums.Viewpoint;
-import com.goormgb.be.ordercore.onboarding.repository.OnboardingPreferenceRepository;
 import com.goormgb.be.user.entity.User;
 import com.goormgb.be.user.repository.UserRepository;
