+++ Order-Core/src/test/java/com/goormgb/be/ordercore/fixture/onboarding/OnboardingPreferenceFixture.java
@@ -2,15 +2,15 @@ package com.goormgb.be.ordercore.fixture.onboarding;
 
 import org.springframework.test.util.ReflectionTestUtils;
 
-import com.goormgb.be.ordercore.onboarding.entity.OnboardingPreference;
-import com.goormgb.be.ordercore.onboarding.enums.EnvironmentPref;
-import com.goormgb.be.ordercore.onboarding.enums.MoodPref;
-import com.goormgb.be.ordercore.onboarding.enums.ObstructionSensitivity;
-import com.goormgb.be.ordercore.onboarding.enums.PriceMode;
-import com.goormgb.be.ordercore.onboarding.enums.SeatHeight;
-import com.goormgb.be.ordercore.onboarding.enums.SeatPositionPref;
-import com.goormgb.be.ordercore.onboarding.enums.Section;
-import com.goormgb.be.ordercore.onboarding.enums.Viewpoint;
+import com.goormgb.be.onboarding.entity.OnboardingPreference;
+import com.goormgb.be.onboarding.enums.EnvironmentPref;
+import com.goormgb.be.onboarding.enums.MoodPref;
+import com.goormgb.be.onboarding.enums.ObstructionSensitivity;
+import com.goormgb.be.onboarding.enums.PriceMode;
+import com.goormgb.be.onboarding.enums.SeatHeight;
+import com.goormgb.be.onboarding.enums.SeatPositionPref;
+import com.goormgb.be.onboarding.enums.Section;
+import com.goormgb.be.onboarding.enums.Viewpoint;
 import com.goormgb.be.user.entity.User;
 
 public final class OnboardingPreferenceFixture {
@@ -24,51 +24,51 @@ public final class OnboardingPreferenceFixture {
 
 	public static OnboardingPreference createFirst(User user) {
 		return OnboardingPreference.builder()
-				.user(user)
-				.priority(1)
-				.viewpoint(Viewpoint.CENTER)
-				.seatHeight(SeatHeight.LOW)
-				.section(Section.CENTER_SIDE)
-				.seatPositionPref(SeatPositionPref.AISLE)
-				.environmentPref(EnvironmentPref.SHADE)
-				.moodPref(MoodPref.CHEERFUL)
-				.obstructionSensitivity(ObstructionSensitivity.NORMAL)
-				.priceMode(PriceMode.RANGE)
-				.priceMin(30000)
-				.priceMax(80000)
-				.build();
+			.user(user)
+			.priority(1)
+			.viewpoint(Viewpoint.CENTER)
+			.seatHeight(SeatHeight.LOW)
+			.section(Section.CENTER_SIDE)
+			.seatPositionPref(SeatPositionPref.AISLE)
+			.environmentPref(EnvironmentPref.SHADE)
+			.moodPref(MoodPref.CHEERFUL)
+			.obstructionSensitivity(ObstructionSensitivity.NORMAL)
+			.priceMode(PriceMode.RANGE)
+			.priceMin(30000)
+			.priceMax(80000)
+			.build();
 	}
 
 	public static OnboardingPreference createSecond(User user) {
 		return OnboardingPreference.builder()
-				.user(user)
-				.priority(2)
-				.viewpoint(Viewpoint.INFIELD_1B)
-				.seatHeight(SeatHeight.MID)
-				.section(Section.MIDDLE)
-				.seatPositionPref(SeatPositionPref.ANY)
-				.environmentPref(EnvironmentPref.SUN_OK)
-				.moodPref(MoodPref.QUIET)
-				.obstructionSensitivity(ObstructionSensitivity.NET_SENSITIVE)
-				.priceMode(PriceMode.ANY)
-				.build();
+			.user(user)
+			.priority(2)
+			.viewpoint(Viewpoint.INFIELD_1B)
+			.seatHeight(SeatHeight.MID)
+			.section(Section.MIDDLE)
+			.seatPositionPref(SeatPositionPref.ANY)
+			.environmentPref(EnvironmentPref.SUN_OK)
+			.moodPref(MoodPref.QUIET)
+			.obstructionSensitivity(ObstructionSensitivity.NET_SENSITIVE)
+			.priceMode(PriceMode.ANY)
+			.build();
 	}
 
 	public static OnboardingPreference createThird(User user) {
 		return OnboardingPreference.builder()
-				.user(user)
-				.priority(3)
-				.viewpoint(Viewpoint.OUTFIELD_L)
-				.seatHeight(SeatHeight.HIGH)
-				.section(Section.CORNER)
-				.seatPositionPref(SeatPositionPref.MIDDLE)
-				.environmentPref(EnvironmentPref.ANY)
-				.moodPref(MoodPref.ANY)
-				.obstructionSensitivity(ObstructionSensitivity.RAIL_PILLAR_SENSITIVE)
-				.priceMode(PriceMode.RANGE)
-				.priceMin(10000)
-				.priceMax(50000)
-				.build();
+			.user(user)
+			.priority(3)
+			.viewpoint(Viewpoint.OUTFIELD_L)
+			.seatHeight(SeatHeight.HIGH)
+			.section(Section.CORNER)
+			.seatPositionPref(SeatPositionPref.MIDDLE)
+			.environmentPref(EnvironmentPref.ANY)
+			.moodPref(MoodPref.ANY)
+			.obstructionSensitivity(ObstructionSensitivity.RAIL_PILLAR_SENSITIVE)
+			.priceMode(PriceMode.RANGE)
+			.priceMin(10000)
+			.priceMax(50000)
+			.build();
 	}
 
 	public static OnboardingPreference createWithId(Long id, User user) {
