+++ common-core/src/main/java/com/goormgb/be/onboarding/enums/EnvironmentPref.java
@@ -1,13 +1,13 @@
-package com.goormgb.be.ordercore.onboarding.enums;
+package com.goormgb.be.onboarding.enums;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 
 @Getter
 @RequiredArgsConstructor
-public enum SeatPositionPref {
-	AISLE("통로석 선호"),
-	MIDDLE("중앙석 선호"),
+public enum EnvironmentPref {
+	SHADE("그늘 선호"),
+	SUN_OK("햇빛 무관"),
 	ANY("무관");
 
 	private final String description;
