+++ common-core/src/main/java/com/goormgb/be/onboarding/enums/MoodPref.java
@@ -1,14 +1,13 @@
-package com.goormgb.be.ordercore.onboarding.enums;
+package com.goormgb.be.onboarding.enums;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 
 @Getter
 @RequiredArgsConstructor
-public enum SeatHeight {
-	LOW("하단"),
-	MID("중단"),
-	HIGH("상단"),
+public enum MoodPref {
+	CHEERFUL("열정적인 응원"),
+	QUIET("조용한 관람"),
 	ANY("무관");
 
 	private final String description;
