+++ common-core/src/main/java/com/goormgb/be/onboarding/enums/SeatHeight.java
@@ -1,13 +1,14 @@
-package com.goormgb.be.ordercore.onboarding.enums;
+package com.goormgb.be.onboarding.enums;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 
 @Getter
 @RequiredArgsConstructor
-public enum EnvironmentPref {
-	SHADE("그늘 선호"),
-	SUN_OK("햇빛 무관"),
+public enum SeatHeight {
+	LOW("하단"),
+	MID("중단"),
+	HIGH("상단"),
 	ANY("무관");
 
 	private final String description;
