+++ common-core/src/main/java/com/goormgb/be/onboarding/enums/SeatPositionPref.java
@@ -1,13 +1,13 @@
-package com.goormgb.be.ordercore.onboarding.enums;
+package com.goormgb.be.onboarding.enums;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 
 @Getter
 @RequiredArgsConstructor
-public enum MoodPref {
-	CHEERFUL("열정적인 응원"),
-	QUIET("조용한 관람"),
+public enum SeatPositionPref {
+	AISLE("통로석 선호"),
+	MIDDLE("중앙석 선호"),
 	ANY("무관");
 
 	private final String description;
